modified example style

+++ public/js/popup/test-popup.jsx
@@ -34,14 +34,18 @@ var TestPopup  = React.createClass({
                <div>
                  <div className="full-content">
                  { this.state.items.map((item, index) => {
-                    return <div className="item">
+                    return <div key={index} className="item">
                       <button onClick={()=>{ this.displayPopup(index)} }>Delete</button>
                       <span>{item}</span>
                     </div>
                   })
                 }
               </div>
-		          <Popup display={this.state.showPopup} onYes={this.deleteItem} onNo={this.onNo}/>
+		          <Popup display={this.state.showPopup} title="Confirmation" onYes={this.deleteItem} onNo={this.onNo}>
+                <div>
+                    Do you want to delete item {this.state.currentItemIndex + 1 }?
+                </div>
+              </Popup>
                 </div>
 		   );
    }
@@ -52,10 +56,13 @@ var Popup = React.createClass({
         return (
                 <div className="popup" style={{display: this.props.display? "block": "none"}}>
                     <div className="popup-container">
-                        <div className="title">Are you sure?</div>
+                        <div className="title">{this.props.title}</div>
                         <div className="content">
-                          <button className="secondary" type="button" onClick={this.props.onYes}>Yes</button>
-                          <button className="primary" type="button" onClick={this.props.onNo}>No</button>
+                          <div>{this.props.children}</div>
+                          <div className="footer">
+                            <button className="secondary" type="button" onClick={this.props.onYes}>Yes</button>
+                            <button className="primary" type="button" onClick={this.props.onNo}>No</button>
+                          </div>
                         </div>
                     </div>
                 </div>
